미션 11 초급 수정
- 변수타입 변경
- 메인 추가

+++ nahye/mission11/basic/Rectangle.java
@@ -2,10 +2,10 @@ package mission11.basic;
 
 public class Rectangle extends Shape {
 
-    private int width;
-    private int height;
+    private double width;
+    private double height;
 
-    public Rectangle(int width, int height) {
+    public Rectangle(double width, double height) {
         this.width = width;
         this.height = height;
     }
